refactor: IntegrationTestSupport 패키지 변경
